[hijax] type-changing boxes with pytree contents
